+++ source/gameshared/q_comref.h
@@ -330,7 +330,7 @@ typedef struct entity_state_s {
 	vec3_t linearMovementEnd;           // the end movement point for brush models
 	vec3_t linearMovementBegin;     // the starting movement point for brush models
 	unsigned int linearMovementDuration;
-	unsigned int linearMovementTimeStamp;
+	int64_t linearMovementTimeStamp;
 
 	float attenuation;                  // should be <= 255/16.0 as this is sent as byte
